project_with_weather_from_csv/requirements.txt
psycopg2-binary
pandas
seaborn
matplotlib
numpy
requests